requirements.txt
tqdm
scikit-image
numpy
deep-pipe
nibabel
pynrrd
torch
